config-management/requirements.txt
redis==5.0.1
python-consul==1.1.0
hvac==2.1.0
pydantic==2.5.3
pyyaml==6.0.1
